s1/requirements.txt
flask
grpcio-tools
imagination
opentelemetry-distro
opentelemetry-exporter-otlp
pydantic
requests
